+++ src/print_signed.c
@@ -6,14 +6,14 @@
 /*   By: smamalig <smamalig@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/06 15:08:04 by smamalig          #+#    #+#             */
-/*   Updated: 2025/05/09 02:43:32 by smamalig         ###   ########.fr       */
+/*   Updated: 2025/05/16 10:48:07 by smamalig         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft_internal.h"
 
 static void	__ft_printf_int_internal(t_printf_parser *parser,
-	__int128_t n, int iters)
+	intmax_t n, int iters)
 {
 	if (!iters)
 		return ;
@@ -21,7 +21,7 @@ static void	__ft_printf_int_internal(t_printf_parser *parser,
 	__ft_printf_insert(parser, __ft_printf_abs_char(n % 10) + '0');
 }
 
-int	__ft_printf_print_len(t_printf_parser *parser, __int128_t n)
+int	__ft_printf_print_len(t_printf_parser *parser, intmax_t n)
 {
 	const int	len = __ft_printf_int_len(n);
 
@@ -32,7 +32,7 @@ int	__ft_printf_print_len(t_printf_parser *parser, __int128_t n)
 	return (len);
 }
 
-void	__ft_printf_int(t_printf_parser *parser, __int128_t n)
+void	__ft_printf_int(t_printf_parser *parser, intmax_t n)
 {
 	const int	len = __ft_printf_print_len(parser, n);
 	const int	sign = n < 0 || parser->flags & PRINTF_FLAG_SIGN;
